simplifying first pass at Server URL parsing to use as variable parsing for any URLTemplate.
