+++ Sources/OpenAPIKit/URLTemplate/URLTemplate.swift
@@ -39,6 +39,16 @@ public struct URLTemplate: Hashable, RawRepresentable {
     /// (i.e. the entire URL).
     public let rawValue: String
 
+    /// The variable and constant components of teh URL.
+    ///
+    /// Variables are enclosed in curly braces (`{variable}`)
+    /// and everything between variables is "constant."
+    ///
+    /// In the URL `{scheme}://website.com/{path}`
+    /// there are two variables (`scheme` and `path`) separated
+    /// by the "://website.com/" portion which is a constant.
+    public let components: [Component]
+
     /// The entire URL as a string.
     ///
     /// This is equivalent to the `absoluteString` provided
@@ -58,16 +68,59 @@ public struct URLTemplate: Hashable, RawRepresentable {
     }
 
     /// Create a URLTemplate from the string if possible.
+    ///
+    /// A non-throwing version of `init(templateString:)`.
     public init?(rawValue: String) {
-        // currently this is guaranteed to succeed but
-        // in the future it will fail if variable placeholders
-        // cannot be successfully parsed (i.e. curly braces
-        // are mismatched, dangling, nested, etc.).
-        self.rawValue = rawValue
+        do {
+            try self.init(templateString: rawValue)
+        } catch {
+            return nil
+        }
     }
 
+    /// Create a URLTemplate from the string if possible.
+    ///
+    /// A throwing version of `init?(rawValue:)`.
+    public init(templateString: String) throws {
+        rawValue = templateString
+        components = try URLTemplate.scan(
+            templateString,
+            partialToken: nil,
+            from: templateString[...],
+            addingTo: []
+        )
+    }
+
+    /// Create a constant (no variables) URLTemplate from the given URL.
     public init(url: URL) {
         rawValue = url.absoluteString
+        components = [.constant(rawValue)]
+    }
+}
+
+extension URLTemplate {
+    public enum Component: Hashable, RawRepresentable {
+        case variable(name: String)
+        case constant(String)
+
+        public var rawValue: String {
+            switch self {
+            case .variable(name: let name):
+                return "{\(name)}"
+            case .constant(let value):
+                return value
+            }
+        }
+
+        public init?(rawValue: String) {
+            let first = rawValue.first
+            let last = rawValue.last
+            guard first == "{" && last == "}" else {
+                self = .constant(rawValue)
+                return
+            }
+            self = .variable(name: String(rawValue.dropFirst().dropLast()))
+        }
     }
 }
 
@@ -83,11 +136,11 @@ extension URLTemplate: Decodable {
     public init(from decoder: Decoder) throws {
         let container = try decoder.singleValueContainer()
 
-        rawValue = try container.decode(String.self)
+        let rawValue = try container.decode(String.self)
 
-        // in the future the plan is to tokenize upon
-        // decoding (and other initialization above) and offer
-        // easy variable replacement.
+        let template = try URLTemplate(templateString: rawValue)
+
+        self = template
     }
 }
