+++ Tests/OpenAPIKitTests/URLTemplate/URLTemplateTests.swift
@@ -13,6 +13,7 @@ final class URLTemplateTests: XCTestCase {
     func test_init() {
         XCTAssertNotNil(URLTemplate(rawValue: "https://website.com"))
         XCTAssertEqual(URLTemplate(rawValue: "https://website.com"), URLTemplate(url: URL(string: "https://website.com")!))
+        XCTAssertNotNil(URLTemplate(rawValue: "{scheme}://website.com"))
     }
 
     func test_urlAccess() {
@@ -28,9 +29,40 @@ final class URLTemplateTests: XCTestCase {
         let t2 = URLTemplate(rawValue: "/a/relative/path")
         let t3 = URLTemplate(rawValue: "website.com?query=value")
 
+        let t4 = URLTemplate(rawValue: "{scheme}://website.com/{path}")
+
         XCTAssertEqual(t1?.absoluteString, URL(string: "https://website.com")?.absoluteString)
         XCTAssertEqual(t2?.absoluteString, URL(string: "/a/relative/path")?.absoluteString)
         XCTAssertEqual(t3?.absoluteString, URL(string: "website.com?query=value")?.absoluteString)
+
+        XCTAssertEqual(t4?.absoluteString, "{scheme}://website.com/{path}")
+    }
+
+    func test_componentErrors() {
+        // unclosed variable brace
+        XCTAssertNil(URLTemplate(rawValue: "{scheme://website.com"))
+        XCTAssertThrowsError(try URLTemplate(templateString: "{scheme://website.com")) { error in
+            XCTAssertEqual(
+                String(describing: error),
+                "An opening brace with no closing brace was found. The portion of the URL following the opening brace was 'scheme://website.com'"
+            )
+        }
+
+        // unopened variable brace
+        XCTAssertNil(URLTemplate(rawValue: "scheme}://website.com"))
+        XCTAssertThrowsError(try URLTemplate(templateString: "scheme}://website.com")) { error in
+            XCTAssertEqual(String(describing: error), "A closing brace with no opening brace was found. The portion of the URL preceeding the closing brace was 'scheme'")
+        }
+
+        // nested variable brace
+        XCTAssertNil(URLTemplate(rawValue: "{scheme}://website.com/{path{var}}"))
+        XCTAssertThrowsError(try URLTemplate(templateString: "{scheme}://website.com/{path{var}}")) { error in
+            XCTAssertEqual(String(describing: error), "An opening brace within another variable was found. The portion of the URL following the first opening brace up until the second opening brace was 'path'")
+        }
+    }
+
+    func test_componentSuccesses() {
+
     }
 }
